Important update to spacebar handling
The space key being tied to silence caused major problems by typing when the user does not wish (or becoming a BADUSB attack), so it's been given it's own keycode. 
The keyboard method needs to be switched from keyboard_layout_us to keyboard_press in order to allow other components of the keyboard such as CTRL and SHIFT which will be toggled.